Lab04 req07

+++ app/src/main/java/ar/edu/utn/frsf/dam/isi/laboratorio02/dao/CategoriaDao.java
@@ -22,5 +22,5 @@ public interface CategoriaDao {
     void delete(Categoria c);
 
     @Query("SELECT * FROM Categoria WHERE id = :idCategoria")
-    Categoria buscarCategoriaPorId(long idCategoria);
+    Categoria buscarCategoriaPorId(int idCategoria);
 }
